Modified config loading to add logging of non-secret values and warnings of skipped files
