+++ dataactcore/config.py
@@ -1,138 +1,5 @@
-import os
-import os.path
-from os.path import expanduser, normpath, dirname, abspath
-import yaml
-import re
+from dataactcore.read_config import CONFIG_BROKER, CONFIG_LOGGING, CONFIG_SERVICES, CONFIG_DB, CONFIG_JOB_QUEUE, log_message
+from dataactcore.utils.cloudLogger import CloudLogger
 
-env = os.environ["env"] if "env" in os.environ else "dev"
-# set the location of the DATA Act broker config file
-CONFIG_PATH = os.path.join(dirname(abspath(__file__)), 'config.yml')
-ENV_PATH = os.path.join(dirname(abspath(__file__)), '{}_config.yml'.format(env))
-# set the location of the Alembic config file
-ALEMBIC_PATH = os.path.join(dirname(abspath(__file__)), 'alembic.ini')
-MIGRATION_PATH = os.path.join(dirname(abspath(__file__)), 'migrations')
-CONFIG_BROKER = {}
-CONFIG_SERVICES = {}
-CONFIG_DB = {}
-CONFIG_LOGGING = {}
-CONFIG_JOB_QUEUE = {}
-
-def merge_config(existing_config, new_config):
-    if type(existing_config) is not dict or type(new_config) is not dict:
-        raise ValueError("Both args to merge_config must be dicts")
-    for key in new_config.keys():
-        existing_config[key] = new_config[key]
-    return existing_config
-
-first_config = True
-for config_path in [CONFIG_PATH, ENV_PATH]:
-    try:
-        with open(config_path) as c:
-            CONFIG_ALL = yaml.load(c)
-    except IOError:
-        if first_config:
-            # First config file must be present, the rest are optional
-            raise IOError('Error reading the config file. Please make sure this file exists'
-               ' before starting the DATA Act broker: {}'.format(config_path))
-        else:
-            continue
-    # File successfully loaded, mark as no longer first
-    first_config = False
-    
-    CONFIG_BROKER = merge_config(CONFIG_BROKER, CONFIG_ALL['broker']) if "broker" in CONFIG_ALL else CONFIG_BROKER
-    CONFIG_SERVICES = merge_config(CONFIG_SERVICES, CONFIG_ALL['services']) if "services" in CONFIG_ALL else CONFIG_SERVICES
-    CONFIG_DB = merge_config(CONFIG_DB, CONFIG_ALL['db']) if "db" in CONFIG_ALL else CONFIG_DB
-    CONFIG_LOGGING = merge_config(CONFIG_LOGGING, CONFIG_ALL['logging']) if "logging" in CONFIG_ALL else CONFIG_LOGGING
-    CONFIG_JOB_QUEUE = merge_config(CONFIG_JOB_QUEUE, CONFIG_ALL['job-queue']) if "job-queue" in CONFIG_ALL else CONFIG_JOB_QUEUE
-
-# Get path to installation
-CONFIG_BROKER['path'] = dirname(dirname(abspath(__file__)))
-
-# for backward-compatibility, differentiate between local runs and AWS
-if CONFIG_BROKER['use_aws'] is True or CONFIG_BROKER['use_aws'] == "true":
-    CONFIG_BROKER['local'] = False
-    # AWS flag is on, so make sure all needed AWS info is present
-    required_aws_keys = ['aws_bucket', 'aws_role', 'aws_region', 'aws_create_temp_credentials',
-                         'static_files_bucket', 'help_files_path']
-    for k in required_aws_keys:
-        try:
-            CONFIG_BROKER[k]
-        except KeyError as e:
-            raise KeyError('Config error: use_aws is True, but the {} key is'
-                ' missing from the config.yml file'.format(k))
-        if not CONFIG_BROKER[k]:
-            raise ValueError('Config error: use_aws is True but {} value is '
-                 'missing'.format(k))
-
-    help_files_path = CONFIG_BROKER["help_files_path"]
-    CONFIG_BROKER["help_files_path"] = "".join([help_files_path, "/"]) if help_files_path[-1] != "/" else help_files_path
-else:
-    CONFIG_BROKER['local'] = True
-    CONFIG_BROKER['aws_bucket'] = None
-    CONFIG_BROKER['aws_role'] = None
-    CONFIG_BROKER['aws_create_temp_credentials'] = None
-    CONFIG_BROKER['aws_region'] = None
-
-    # if not using AWS and no error report path specified,
-    # default to `data_act_broker` in user's home dir
-    error_report_path = CONFIG_SERVICES['error_report_path']
-    if not error_report_path:
-        error_report_path = os.path.join(expanduser('~'), 'data_act_broker')
-    normpath(error_report_path)
-    CONFIG_SERVICES['error_report_path'] = error_report_path
-
-    # if not using AWS and no DynamoDB settings specified, default
-    # to localhost and port 8000
-    if not CONFIG_DB['dynamo_host']:
-        # TODO: dynamo host/port don't do anything; code assumes localhost/8000
-        CONFIG_DB['dynamo_host'] = '127.0.0.1'
-    if not CONFIG_DB['dynamo_port']:
-        CONFIG_DB['dynamo_port'] = 8000
-    # TODO: can we test that local dynamo is up and running? if not, route calls hang
-
-storage_path = CONFIG_BROKER['d_file_storage_path']
-if storage_path[-1] != os.path.sep:
-    CONFIG_BROKER['d_file_storage_path'] = "".join([storage_path, os.path.sep])
-
-# if no broker file path specified,
-# default to `data_act_broker` in user's home dir
-broker_files = CONFIG_BROKER['broker_files']
-if not broker_files:
-    broker_files = os.path.join(expanduser('~'), 'data_act_broker')
-elif len(os.path.splitext(broker_files)[1]):
-    # if config's broker_files is set to a actual filename
-    # just use the directory
-    broker_files = os.path.split(broker_files)[0]
-normpath(broker_files)
-if broker_files[-1] != os.path.sep:
-    broker_files += os.path.sep
-CONFIG_BROKER['broker_files'] = broker_files
-
-# normalize logging path, if given
-log_path = CONFIG_LOGGING['log_files']
-if log_path:
-    CONFIG_LOGGING['log_files'] = normpath(log_path)
-
-# we don't want http:// or ports in the host variables
-CONFIG_SERVICES['broker_api_host'] = re.sub(
-    'http://|:(.*)', '', CONFIG_SERVICES['broker_api_host'])
-CONFIG_SERVICES['validator_host'] = re.sub(
-    'http://|:(.*)', '', CONFIG_SERVICES['validator_host'])
-# if hosts in config file are set to 0.0.0.0, override to
-# 127.0.0.1 for cross-platform compatibility
-if CONFIG_SERVICES['broker_api_host'] == '0.0.0.0':
-    CONFIG_SERVICES['broker_api_host'] = '127.0.0.1'
-if CONFIG_SERVICES['validator_host'] == '0.0.0.0':
-    CONFIG_SERVICES['validator_host'] = '127.0.0.1'
-if CONFIG_DB['dynamo_host'] == '0.0.0.0':
-    CONFIG_DB['dynamo_host'] = '127.0.0.1'
-
-if CONFIG_SERVICES["broker_api_port"] == 443:
-    # Use https
-    CONFIG_SERVICES["protocol"] = "https"
-else:
-    CONFIG_SERVICES["protocol"] = "http"
-
-# TODO: error-handling for db config?
-# TODO: type checking and fixing for int stuff like ports?
-
+# Log config values along with warnings for missing files
+CloudLogger.log(log_message)
